refactor code and support additional csv data for FastAPI
