+++ templates/types/streaming/fastapi/app/api/routers/messaging.py
@@ -1,149 +1,170 @@
-import json
-import asyncio
-import logging
-from typing import AsyncGenerator, Dict, Any, List, Optional
-from llama_index.core.callbacks.base import BaseCallbackHandler
-from llama_index.core.callbacks.schema import CBEventType
-from llama_index.core.tools.types import ToolOutput
-from pydantic import BaseModel
+from pydantic import BaseModel, Field, validator
+from pydantic.alias_generators import to_camel
+from typing import List, Any, Optional, Dict
+from llama_index.core.schema import NodeWithScore
+from llama_index.core.llms import ChatMessage, MessageRole
 
 
-logger = logging.getLogger(__name__)
+class Message(BaseModel):
+    role: MessageRole
+    content: str
 
 
-class CallbackEvent(BaseModel):
-    event_type: CBEventType
-    payload: Optional[Dict[str, Any]] = None
-    event_id: str = ""
+class CsvFile(BaseModel):
+    content: str
+    filename: str
+    filesize: int
+    id: str
+    type: str
 
-    def get_retrieval_message(self) -> dict | None:
-        if self.payload:
-            nodes = self.payload.get("nodes")
-            if nodes:
-                msg = f"Retrieved {len(nodes)} sources to use as context for the query"
-            else:
-                msg = f"Retrieving context for query: '{self.payload.get('query_str')}'"
-            return {
-                "type": "events",
-                "data": {"title": msg},
-            }
-        else:
-            return None
 
-    def get_tool_message(self) -> dict | None:
-        func_call_args = self.payload.get("function_call")
-        if func_call_args is not None and "tool" in self.payload:
-            tool = self.payload.get("tool")
-            return {
-                "type": "events",
-                "data": {
-                    "title": f"Calling tool: {tool.name} with inputs: {func_call_args}",
-                },
-            }
+class DataParserOptions(BaseModel):
+    csv_files: List[CsvFile] | None = Field(
+        default=None,
+        description="List of CSV files",
+    )
 
-    def _is_output_serializable(self, output: Any) -> bool:
-        try:
-            json.dumps(output)
-            return True
-        except TypeError:
-            return False
-
-    def get_agent_tool_response(self) -> dict | None:
-        response = self.payload.get("response")
-        if response is not None:
-            sources = response.sources
-            for source in sources:
-                # Return the tool response here to include the toolCall information
-                if isinstance(source, ToolOutput):
-                    if self._is_output_serializable(source.raw_output):
-                        output = source.raw_output
-                    else:
-                        output = source.content
-
-                    return {
-                        "type": "tools",
-                        "data": {
-                            "toolOutput": {
-                                "output": output,
-                                "isError": source.is_error,
-                            },
-                            "toolCall": {
-                                "id": None,  # There is no tool id in the ToolOutput
-                                "name": source.tool_name,
-                                "input": source.raw_input,
-                            },
-                        },
+    class Config:
+        json_schema_extra = {
+            "example": {
+                "csvFiles": [
+                    {
+                        "content": "Name, Age\nAlice, 25\nBob, 30",
+                        "filename": "example.csv",
+                        "filesize": 123,
+                        "id": "123",
+                        "type": "text/csv",
                     }
+                ]
+            }
+        }
+        alias_generator = to_camel
 
-    def to_response(self):
-        try:
-            match self.event_type:
-                case "retrieve":
-                    return self.get_retrieval_message()
-                case "function_call":
-                    return self.get_tool_message()
-                case "agent_step":
-                    return self.get_agent_tool_response()
-                case _:
-                    return None
-        except Exception as e:
-            logger.error(f"Error in converting event to response: {e}")
-            return None
+    def to_raw_content(self) -> str:
+        if self.csv_files is not None and len(self.csv_files) > 0:
+            return "Use data from following CSV raw contents" + "\n".join(
+                [f"```csv\n{csv_file.content}\n```" for csv_file in self.csv_files]
+            )
+
+    def to_response_data(self) -> list[dict] | None:
+        output = []
+        if self.csv_files is not None and len(self.csv_files) > 0:
+            output.append(
+                {
+                    "type": "csv",
+                    "data": {
+                        "csvFiles": [csv_file.dict() for csv_file in self.csv_files]
+                    },
+                }
+            )
+        return output if len(output) > 0 else None
 
 
-class EventCallbackHandler(BaseCallbackHandler):
-    _aqueue: asyncio.Queue
-    is_done: bool = False
+class ChatData(BaseModel):
+    data: DataParserOptions | None = Field(
+        default=None,
+    )
+    messages: List[Message]
 
-    def __init__(
-        self,
-    ):
-        """Initialize the base callback handler."""
-        ignored_events = [
-            CBEventType.CHUNKING,
-            CBEventType.NODE_PARSING,
-            CBEventType.EMBEDDING,
-            CBEventType.LLM,
-            CBEventType.TEMPLATING,
+    class Config:
+        json_schema_extra = {
+            "example": {
+                "messages": [
+                    {
+                        "role": "user",
+                        "content": "What standards for letters exist?",
+                    }
+                ]
+            }
+        }
+
+    @validator("messages")
+    def messages_must_not_be_empty(cls, v):
+        if len(v) == 0:
+            raise ValueError("Messages must not be empty")
+        return v
+
+    def get_last_message_content(self) -> str:
+        """
+        Get the content of the last message along with the data content if available
+        """
+        message_content = self.messages[-1].content
+        if self.data:
+            message_content += "\n" + self.data.to_raw_content()
+        return message_content
+
+    def get_history_messages(self) -> List[Message]:
+        """
+        Get the history messages
+        """
+        return [
+            ChatMessage(role=message.role, content=message.content)
+            for message in self.messages[:-1]
         ]
-        super().__init__(ignored_events, ignored_events)
-        self._aqueue = asyncio.Queue()
 
-    def on_event_start(
-        self,
-        event_type: CBEventType,
-        payload: Optional[Dict[str, Any]] = None,
-        event_id: str = "",
-        **kwargs: Any,
-    ) -> str:
-        event = CallbackEvent(event_id=event_id, event_type=event_type, payload=payload)
-        if event.to_response() is not None:
-            self._aqueue.put_nowait(event)
+    def get_additional_data_response(self) -> list[dict] | None:
+        """
+        Get the additional data
+        """
+        return self.data.to_response_data()
 
-    def on_event_end(
-        self,
-        event_type: CBEventType,
-        payload: Optional[Dict[str, Any]] = None,
-        event_id: str = "",
-        **kwargs: Any,
-    ) -> None:
-        event = CallbackEvent(event_id=event_id, event_type=event_type, payload=payload)
-        if event.to_response() is not None:
-            self._aqueue.put_nowait(event)
+    def is_last_message_from_user(self) -> bool:
+        return self.messages[-1].role == MessageRole.USER
 
-    def start_trace(self, trace_id: Optional[str] = None) -> None:
-        """No-op."""
 
-    def end_trace(
-        self,
-        trace_id: Optional[str] = None,
-        trace_map: Optional[Dict[str, List[str]]] = None,
-    ) -> None:
-        """No-op."""
+class SourceNodes(BaseModel):
+    id: str
+    metadata: Dict[str, Any]
+    score: Optional[float]
+    text: str
+    url: Optional[str]
 
-    async def async_event_gen(self) -> AsyncGenerator[CallbackEvent, None]:
-        while not self._aqueue.empty() or not self.is_done:
-            try:
-                yield await asyncio.wait_for(self._aqueue.get(), timeout=0.1)
-            except asyncio.TimeoutError:
-                pass
+    @classmethod
+    def from_source_node(cls, source_node: NodeWithScore):
+        metadata = source_node.node.metadata
+        url = metadata.get("URL")
+
+        if not url:
+            file_name = metadata.get("file_name")
+            url_prefix = os.getenv("FILESERVER_URL_PREFIX")
+            if not url_prefix:
+                logger.warning(
+                    "Warning: FILESERVER_URL_PREFIX not set in environment variables"
+                )
+            if file_name and url_prefix:
+                url = f"{url_prefix}/data/{file_name}"
+
+        return cls(
+            id=source_node.node.node_id,
+            metadata=metadata,
+            score=source_node.score,
+            text=source_node.node.text,  # type: ignore
+            url=url,
+        )
+
+    @classmethod
+    def from_source_nodes(cls, source_nodes: List[NodeWithScore]):
+        return [cls.from_source_node(node) for node in source_nodes]
+
+
+class Result(BaseModel):
+    result: Message
+    nodes: List[SourceNodes]
+
+
+class ChatConfig(BaseModel):
+    starter_questions: Optional[List[str]] = Field(
+        default=None,
+        description="List of starter questions",
+    )
+
+    class Config:
+        json_schema_extra = {
+            "example": {
+                "starterQuestions": [
+                    "What standards for letters exist?",
+                    "What are the requirements for a letter to be considered a letter?",
+                ]
+            }
+        }
+        alias_generator = to_camel
